+++ ObjectMapperAdditions/Classes/Realm/RealmTransform.swift
@@ -11,35 +11,26 @@ import RealmAdditions
 import ObjectMapper
 
 
-/// Transforms Swift Arrays to Realm Arrays. E.g. [String] to List<RealmString>.
-/// It supports Int, Double, Bool and String types.
-public class RealmTransform<T: Object>: TransformType where T: RealmValue {
+/// Transforms Swift Arrays to Realm Arrays. E.g. [String] to List<String>.
+public class RealmTransform<T: RealmCollectionValue>: TransformType {
     public typealias Object = List<T>
     public typealias JSON = [Any]
     
     public init() {}
     
-    public func transformFromJSON(_ value: Any?) -> Object? {
+    public func transformFromJSON(_ value: Any?) -> List<T>? {
         guard let array = value as? [Any] else { return nil }
         
-        let realmValues: [T] = array.flatMap({
-            guard let value = $0 as? T.ValueType else { return nil }
-            
-            let realmValue = T()
-            realmValue.value = value
-            
-            return realmValue
-        })
-        
         let list = List<T>()
+        let realmValues: [T] = array.flatMap { $0 as? T }
         list.append(objectsIn: realmValues)
         
         return list
     }
     
-    public func transformToJSON(_ value: Object?) -> JSON? {
+    public func transformToJSON(_ value: List<T>?) -> [Any]? {
         guard let value = value else { return nil }
         
-        return value.map({ $0.value })
+        return Array(value)
     }
 }